Use hostname in endpoint rather than `--hostname` flag
This is because although `${{ github.api_url }}` is "https://api.github.com", the `--hostname` flag expects a base domain, e.g. `github.com` that the "https://api." prefix is appended to internally. Providing the hostname in the form of `HOSTNAME/ENDPOINT` naturally avoids this as the path will infer the `https://api.[github.com]/` prefix to the endpoint, if provided explicitly it becomes no longer necessary to specify the `--hostname` flag.